+++ CHANGELOG.md
@@ -57,6 +57,7 @@
 - **BREAKING**: `ImageCacheKey` is no longer an enum.
   - Switching to a struct allows some optimizations for faster `Hash` and `PartialEq` implementations.
 - **BREAKING**: User cache keys provided via the `ImageCacheKey::user` method now additionally require `Send + Sync`.
+- **BREAKING**: `ImageCacheKey::path` now takes `AsRef<Path>` instead of `Into<String>`. This shouldn't break most cases.
 - Added `is_` and `as_` methods to `ImageCacheKey` for each variant to replace functionality lost.
   - Notably this allows the user key to be retrived which previously wasn't possible.
 - Added `ImageCache::load_from_cache_key` method to `ImageCache`.